// src/app/core/models/model/userResponse.ts
/**
 * Wrapity
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserResponse { 
    id: number;
    username: string;
    email: string;
    bio?: string | null;
    avatar_url?: string | null;
    role: string;
}