// src/app/core/models/model/userUpdate.ts
/**
 * Wrapity
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserUpdate { 
    username?: string | null;
    bio?: string | null;
    avatar_url?: string | null;
}